Fix typo spotted by Dorth.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@2099 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COpenGLExtensionHandler.cpp
@@ -410,7 +410,7 @@ void COpenGLExtensionHandler::initExtensions(bool stencilBuffer)
 	}
 #endif
 	glGetIntegerv(GL_MAX_TEXTURE_SIZE, &num);
-	MaxTextureSize=static_cast<u8>(num);
+	MaxTextureSize=static_cast<u32>(num);
 	glGetIntegerv(GL_MAX_CLIP_PLANES, &num);
 	MaxUserClipPlanes=static_cast<u8>(num);
 	glGetIntegerv(GL_AUX_BUFFERS, &num);
